requirements.txt
colorama==0.4.5
df-engine==0.10.1
python-dotenv==0.20.0
df_runner==0.2.0
git+https://github.com/EvgeniyEgoorov/dialog_flow_db_connector.git@31dd583
ydb
nest_asyncio
asyncio
requests